Move over to `parry3d`, upgrade `nalgebra`

+++ src/resource/vertex_index.rs
@@ -4,7 +4,7 @@ pub use inner::*;
 
 mod inner {
     /// Defaults to `u16`. If you need more than 65535 vertices, enable the `vertex_index_u32` feature.
-    pub type VertexIndex = u16;
+    pub type VertexIndex = u32;
     /// Tells the glow Context what type is used as the vertex index.
     pub const VERTEX_INDEX_TYPE: u32 = crate::context::Context::UNSIGNED_SHORT;
 }
